Support typescript 5.0 with ESM... maybe?
