+++ package.json
@@ -4,11 +4,12 @@
   "description": "Check Digit standard Typescript configuration",
   "prettier": "@checkdigit/prettier-config",
   "engines": {
-    "node": ">=16"
+    "node": ">=18"
   },
+  "type": "module",
   "peerDependencies": {
-    "@types/node": ">=16",
-    "typescript": ">=4.9.3 <5.0"
+    "@types/node": ">=18",
+    "typescript": "^5.0.0-beta"
   },
   "repository": {
     "type": "git",
@@ -27,12 +28,14 @@
     "prettier:fix": "prettier --write .",
     "test": "npm run ci:compile && npm run ci:test && npm run ci:style",
     "ci:compile": "tsc --noEmit",
-    "ci:test": "tsc && node build/index.js | grep -q 'complete' && rimraf build",
+    "ci:test": "tsc && swc build -d swc-build && node swc-build/index.js 2>&1 | grep -q 'complete' && rimraf build",
     "ci:style": "npm run prettier"
   },
   "devDependencies": {
-    "@checkdigit/prettier-config": "^3.1.0",
-    "rimraf": "^3.0.2"
+    "@checkdigit/prettier-config": "^3.2.0",
+    "@swc/cli": "^0.1.60",
+    "@swc/core": "^1.3.32",
+    "rimraf": "^4.1.2"
   },
   "files": [
     "tsconfig.json",
